[1-4] refactor: 패키지 경로 정리
