+++ src/main/java/StringAddCalculator/StringAddCalculator.java
@@ -1,4 +1,4 @@
-package carRacing;
+package StringAddCalculator;
 
 import java.util.Arrays;
 import java.util.regex.Matcher;
