+++ src/test/java/setTest/SetTest.java
@@ -1,4 +1,4 @@
-package carRacing;
+package setTest;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
