+++ src/test/java/stringTest/StringTest.java
@@ -1,4 +1,4 @@
-package carRacing;
+package stringTest;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
